requirements.txt
dashscope==1.20.12
GPUtil==1.4.0
GPUtil==1.4.0
moviepy==1.0.3
numpy==2.1.3
openai==1.54.3
opencv_python==4.10.0.84
oss2==2.19.1
psutil==5.9.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scipy==1.14.1
websocket_client==1.8.0
